The merge wasn't clean, because of changes in build system. The most
significant change was partial revert of usbhc callback refactoring,
which now does not take usb transfer batch, but few named fields again.

+++ kernel/arch/ia32/src/interrupt.c
@@ -87,7 +87,7 @@ void istate_decode(istate_t *istate)
 	    "ebp=%0#10" PRIx32 "\tesp=%0#10" PRIxn "\n",
 	    istate->esi, istate->edi, istate->ebp,
 	    istate_from_uspace(istate) ? istate->esp :
-	    (uintptr_t) &istate->esp);
+	    (uint32_t) &istate->esp);
 }
 
 static void trap_virtual_eoi(void)
